requirements.txt
Beautifulsoup4 == 4.12.3
requests == 2.32.3
pandas == 2.1.4
python-dotenv == 1.0.1